raspberry-pi/requirements.txt
mediacloud==3.12.3
python-dotenv==0.19.0
RPi.GPIO==0.7.0
